+++ src/store/setup.ts
@@ -74,6 +74,10 @@ const driver = createDriver<LegacyStorageSchema & { songs: { key: string; value:
 	name: "beat-mapper-state",
 	version: 3,
 	async upgrade(idb, current, next, tx) {
+		if (current !== 0) {
+			window.alert(`The local database has been updated to version ${next}. Please refresh the page to ensure these migrations can be properly applied.`);
+		}
+
 		// this is a remnant of localforage, and is no longer necessary since blobs are universally supported
 		await idb.removeStore("local-forage-detect-blob-support", tx);
 
@@ -108,7 +112,7 @@ const driver = createDriver<LegacyStorageSchema & { songs: { key: string; value:
 									};
 									return acc;
 								},
-								{} as Record<string, any>,
+								{} as Record<string, unknown>,
 							),
 						},
 						tx,
